[code] fix size_t to int cast to comply to openssl signature EVP_DecryptUpdate EVP_CIPHER_CTX_ctrl EVP_EncryptUpdate

+++ crypto/cipher/aes_icm_ossl.c
@@ -316,7 +316,7 @@ static srtp_err_status_t srtp_aes_icm_openssl_encrypt(void *cv,
         return srtp_err_status_buffer_small;
     }
 
-    if (!EVP_EncryptUpdate(c->ctx, dst, &len, src, src_len)) {
+    if (!EVP_EncryptUpdate(c->ctx, dst, &len, src, (int)src_len)) {
         return srtp_err_status_cipher_fail;
     }
     *dst_len = len;
